requirements.txt
numpy==1.26.4
viam-sdk==0.14.0
